Kahn Academy Animal Attack project
This is the Animal Attack project. It is an animation of a Pikachu that attacks(widens.) I understand the concept of animating and the sizing, but I didn't do the portions for width or hight, just the directions. this way it can do the cool attack. The code is found on Kahn academy into to JS, in the project "animal attack". I am learning Java script